+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ## [1.45.8] - 2026-02-17
 
+### Added
+- post-oauth verify command
+
 ### Bug Fixes
 
 - add jira/linear/status to global AI config command lists (#217)
-
 
 ## [1.45.7] - 2026-02-17